requirements.txt
# API & Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
websockets>=11.0.3

# Agent & LLM
langchain>=0.0.325
langchain-openai>=0.0.2
langchain-community>=0.0.13
autogen>=0.2.0
openai>=1.2.4
google-generativeai>=0.3.1

# RAG & Vector Databases
chromadb>=0.4.18
weaviate-client>=3.24.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pymupdf>=1.23.6
unstructured>=0.10.30
pypdf>=3.17.0

# Vision & OCR
easyocr>=1.7.1
pytesseract>=0.3.10
pillow>=10.1.0
opencv-python>=4.8.1.78
transformers>=4.35.2

# Audio
speechrecognition>=3.10.0
pydub>=0.25.1
elevenlabs>=0.2.24

# Web Crawling & Search
scrapy>=2.11.0
playwright>=1.39.0
google-api-python-client>=2.108.0
serpapi>=0.1.0
beautifulsoup4>=4.12.2
celery>=5.3.4
redis>=5.0.1

# Utility
numpy>=1.24.3
pandas>=2.1.3
tqdm>=4.66.1
loguru>=0.7.2
tenacity>=8.2.3
pytz>=2023.3.post1
aiohttp>=3.8.6
requests>=2.31.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
